updated test runner to operate on all sources
* renamed `tests`->`acceptance-tests`
* fixed how `replace` works to match machine
* add checks for whether acceptance tests should run or not
* add checks for existence of input and expected before running individual regex test
