+++ sources/cz/countrywide.json
@@ -58,7 +58,7 @@
     "test": {
         "enabled": true,
         "description": "these are the tests that exercise valid inputs for czech countrywide",
-        "tests": [
+        "acceptance-tests": [
             {
                 "description": "address with /-delimited number",
                 "inputs": {
